Add Scripting feature (#614)
* update gradle

* update API references to use Long over Int

* cleanup code

* Add in scripting logic

* add in scripting wiki entry

* update keywords to match wiki

* implement mob modifier function

* fixing typos

* registry-aware script update.

+++ src/main/java/harmonised/pmmo/client/gui/TutorialOverlayGUI.java
@@ -149,7 +149,7 @@ public class TutorialOverlayGUI implements LayeredDraw.Layer {
 			SalvageData data = entry.getValue();
 			if (!data.levelReq().isEmpty()) {
 				outList.add(LangProvider.SALVAGE_LEVEL_REQ.asComponent().withStyle(ChatFormatting.UNDERLINE));
-				for (Map.Entry<String, Integer> req : data.levelReq().entrySet()) {
+				for (Map.Entry<String, Long> req : data.levelReq().entrySet()) {
 					outList.add(Component.translatable("pmmo." + req.getKey())
 							.append(Component.literal(": " + req.getValue())));
 				}
